requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
openai>=1.0.0
google-genai>=1.0.0
seaborn>=0.13.0
matplotlib>=3.7.0
